+++ i18n-build/src/gettext_impl/mod.rs
@@ -3,7 +3,7 @@
 
 use crate::error::{PathError, PathType};
 use crate::util;
-use i18n_config::{Crate, GettextConfig, I18nConfig, I18nConfigError};
+use i18n_config::{Crate, GettextConfig, I18nConfigError};
 
 use std::ffi::OsStr;
 use std::fs::{create_dir_all, File};
@@ -226,7 +226,6 @@ pub fn run_msgcat<P: AsRef<Path>, I: IntoIterator<Item = P>>(
 /// stored.
 pub fn run_msginit(
     crt: &Crate,
-    i18n_config: &I18nConfig,
     pot_dir: &Path,
     po_dir: &Path,
 ) -> Result<()> {
@@ -242,7 +241,10 @@ pub fn run_msginit(
 
     let msginit_command_name = "msginit";
 
-    for locale in &i18n_config.target_locales {
+    let gettext_config = crt.gettext_config_or_err()?;
+    let target_locales = &gettext_config.target_locales;
+
+    for locale in target_locales {
         let po_locale_dir = po_dir.join(locale.clone());
         let po_path = po_locale_dir.join(crt.module_name()).with_extension("po");
 
@@ -288,7 +290,6 @@ pub fn run_msginit(
 /// `po_dir` is the directory where the `po` files are stored.
 pub fn run_msgmerge(
     crt: &Crate,
-    i18n_config: &I18nConfig,
     pot_dir: &Path,
     po_dir: &Path,
 ) -> Result<()> {
@@ -302,7 +303,10 @@ pub fn run_msgmerge(
 
     let msgmerge_command_name = "msgmerge";
 
-    for locale in &i18n_config.target_locales {
+    let gettext_config = crt.gettext_config_or_err()?;
+    let target_locales = &gettext_config.target_locales;
+
+    for locale in target_locales {
         let po_file_path = po_dir
             .join(locale)
             .join(crt.module_name())
@@ -337,7 +341,6 @@ pub fn run_msgmerge(
 /// `mo_dir` is the directory where the output `mo` files will be stored.
 pub fn run_msgfmt(
     crt: &Crate,
-    i18n_config: &I18nConfig,
     po_dir: &Path,
     mo_dir: &Path,
 ) -> Result<()> {
@@ -347,7 +350,10 @@ pub fn run_msgfmt(
     );
     let msgfmt_command_name = "msgfmt";
 
-    for locale in &i18n_config.target_locales {
+    let gettext_config = crt.gettext_config_or_err()?;
+    let target_locales = &gettext_config.target_locales;
+
+    for locale in target_locales {
         let po_file_path = po_dir
             .join(locale.clone())
             .join(crt.module_name())
@@ -392,7 +398,7 @@ pub fn run<'a>(crt: &'a Crate) -> Result<()> {
         "Localizing crate \"{0}\" using the gettext system",
         crt.path.to_string_lossy()
     );
-    let (config_crate, i18n_config) = crt.active_config()?.unwrap_or_else(|| {
+    let (config_crate, _i18n_config) = crt.active_config()?.unwrap_or_else(|| {
         panic!(
             "expected that there would be an active config for the crate: \"{0}\" at \"{1}\"",
             crt.name,
@@ -495,9 +501,9 @@ pub fn run<'a>(crt: &'a Crate) -> Result<()> {
     }
 
     if !(crt.collated_subcrate()) {
-        run_msginit(crt, i18n_config, pot_dir.as_path(), po_dir.as_path())?;
-        run_msgmerge(crt, i18n_config, pot_dir.as_path(), po_dir.as_path())?;
-        run_msgfmt(crt, i18n_config, po_dir.as_path(), mo_dir.as_path())?;
+        run_msginit(crt, pot_dir.as_path(), po_dir.as_path())?;
+        run_msgmerge(crt, pot_dir.as_path(), po_dir.as_path())?;
+        run_msgfmt(crt, po_dir.as_path(), mo_dir.as_path())?;
     }
 
     Ok(())