+++ i18n-config/src/lib.rs
@@ -2,6 +2,12 @@
 //! parsing functions) for the
 //! [cargo-i18n](https://crates.io/crates/cargo_i18n) tool/system.
 
+mod gettext;
+mod fluent;
+
+pub use gettext::GettextConfig;
+pub use fluent::FluentConfig;
+
 use std::fs::read_to_string;
 use std::io;
 use std::{
@@ -341,9 +347,7 @@ impl<'a> Display for Crate<'a> {
 #[derive(Deserialize, Debug, Clone)]
 pub struct I18nConfig {
     /// The locale/language identifier of the language used in the source code.
-    pub src_locale: String,
-    /// The locales that the software will be translated into.
-    pub target_locales: Vec<String>,
+    pub fallback_locale: String,
     /// Specify which subcrates to perform localization within. The
     /// subcrate needs to have its own `i18n.toml`.
     #[serde(default)]
@@ -351,6 +355,9 @@ pub struct I18nConfig {
     /// The subcomponent of this config relating to gettext, only
     /// present if the gettext localization system will be used.
     pub gettext: Option<GettextConfig>,
+    /// The subcomponent of this config relating to fluent, only 
+    /// present if the fluent localization system will be used.
+    pub fluent: Option<FluentConfig>,
 }
 
 impl I18nConfig {
@@ -363,120 +370,5 @@ impl I18nConfig {
             I18nConfigError::CannotDeserializeToml(toml_path_final.to_path_buf(), err)
         })?;
         Ok(config)
-    }
-}
-
-#[derive(Deserialize, Debug, Clone)]
-#[serde(rename_all = "lowercase")]
-pub enum GettextAddLocation {
-    Full,
-    File,
-    Never,
-}
-
-impl GettextAddLocation {
-    pub fn to_str(&self) -> &str {
-        match self {
-            GettextAddLocation::Full => "full",
-            GettextAddLocation::File => "file",
-            GettextAddLocation::Never => "never",
-        }
-    }
-}
-
-impl Default for GettextAddLocation {
-    fn default() -> Self {
-        GettextAddLocation::Full
-    }
-}
-
-/// The data structure representing what is stored (and possible to
-/// store) within the `gettext` subsection of a `i18n.toml` file.
-#[derive(Deserialize, Debug, Clone)]
-pub struct GettextConfig {
-    /// Path to the output directory, relative to `i18n.toml` of the
-    /// crate being localized.
-    pub output_dir: PathBuf,
-    // If this crate is being localized as a subcrate, store the
-    // localization artifacts with the parent crate's output.
-    // Currently crates which contain subcrates with duplicate names
-    // are not supported.
-    //
-    // By default this is **false**.
-    #[serde(default)]
-    pub extract_to_parent: bool,
-    // If a subcrate has extract_to_parent set to true,
-    // then merge the output pot file of that subcrate into this
-    // crate's pot file.
-    //
-    // By default this is **false**.
-    #[serde(default)]
-    pub collate_extracted_subcrates: bool,
-    /// Set the copyright holder for the generated files.
-    pub copyright_holder: Option<String>,
-    /// The reporting address for msgid bugs. This is the email
-    /// address or URL to which the translators shall report bugs in
-    /// the untranslated strings.
-    pub msgid_bugs_address: Option<String>,
-    /// Whether or not to perform string extraction using the `xtr` command.
-    pub xtr: Option<bool>,
-    /// Generate ‘#: filename:line’ lines (default) in the pot files when
-    /// running the `xtr` command. If the type is ‘full’ (the default),
-    /// it generates the lines with both file name and line number.
-    /// If it is ‘file’, the line number part is omitted. If it is ‘never’,
-    ///  nothing is generated. [possible values: full, file, never].
-    #[serde(default)]
-    pub add_location: GettextAddLocation,
-    /// Path to where the pot files will be written to by the `xtr`
-    /// command, and were they will be read from by `msginit` and
-    /// `msgmerge`.
-    pot_dir: Option<PathBuf>,
-    /// Path to where the po files will be stored/edited with the
-    /// `msgmerge` and `msginit` commands, and where they will be read
-    /// from with the `msgfmt` command.
-    po_dir: Option<PathBuf>,
-    /// Path to where the mo files will be written to by the
-    /// `msgfmt` command.
-    mo_dir: Option<PathBuf>,
-}
-
-impl GettextConfig {
-    /// Path to where the pot files will be written to by the `xtr`
-    /// command, and were they will be read from by `msginit` and
-    /// `msgmerge`.
-    ///
-    /// By default this is
-    /// **[output_dir](GettextConfig::output_dir)/pot**.
-    pub fn pot_dir(&self) -> PathBuf {
-        // match self.pot_dir {
-        //     Some(pot_dir) => pot_dir,
-        //     None => {
-        //         panic!("panic");
-        //     },
-        // }
-        self.pot_dir
-            .clone()
-            .unwrap_or_else(|| self.output_dir.join("pot"))
-    }
-
-    /// Path to where the po files will be stored/edited with the
-    /// `msgmerge` and `msginit` commands, and where they will
-    /// be read from with the `msgfmt` command.
-    ///
-    /// By default this is **[output_dir](GettextConfig::output_dir)/po**.
-    pub fn po_dir(&self) -> PathBuf {
-        self.po_dir
-            .clone()
-            .unwrap_or_else(|| self.output_dir.join("po"))
-    }
-
-    /// Path to where the mo files will be written to by the `msgfmt` command.
-    ///
-    /// By default this is
-    /// **[output_dir](GettextConfig::output_dir)/mo**.
-    pub fn mo_dir(&self) -> PathBuf {
-        self.mo_dir
-            .clone()
-            .unwrap_or_else(|| self.output_dir.join("mo"))
     }
 }