+++ i18n-build/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog for `i18n-build`
 
+## v0.5.0
+
+Changes for the support of the `fluent` localization system.
+
+### Breaking Changes
+
++ Update to `i18n-embed` version `0.7.0`, contains breaking changes to API.
++ Update to `i18n-config` version `0.3.0`, contains breaking changes to `i18n.toml` configuration file format. See the [i18n changelog](https://github.com/kellpossible/cargo-i18n/blob/master/i18n-config/CHANGELOG.md#v030) for more details.
+
 ## v0.4.1
 
 + Update to `i18n-embed` version `0.6.0`.